requirements.txt
Flask
face_recognition
opencv-python
mysql-connector-python
numpy
cmake
dlib